+++ Configuration/MobileConfiguration.cpp
@@ -60,6 +60,15 @@ void CMobileConfiguration::Clear(){
 	m_maxColumn							= 200.0;
 	m_offsetFrom						= 50;
 	m_offsetTo							= 200;
+
+	// Directory
+	m_directory.Format("");
+	m_sleep = 500;
+	m_spectrometerDyanmicRange = 4096;
+	m_defaultSkyFile.Format("");
+	m_defaultDarkFile.Format("");
+	m_defaultDarkcurFile.Format("");
+	m_defaultOffsetFile.Format("");
 }
 
 /** Reading in a configuration file in .xml file format */
@@ -105,6 +114,10 @@ int CMobileConfiguration::Parse(){
 			// If the serial-port begins with 'COM' then we're using the RS232-port
 			if(Equals(m_serialPort, "COM", 3)){
 				m_spectrometerConnection = CONNECTION_RS232;
+			}
+			// Directory mode enabled
+			if (Equals(m_serialPort, "Directory", 3)) {
+				m_spectrometerConnection = CONNECTION_DIRECTORY;
 			}
 			continue;
 		}
@@ -163,9 +176,21 @@ int CMobileConfiguration::Parse(){
 			continue;
 		}
 
+		// Whether to skip dark measurement
+		if (Equals(szToken, "noDark")) {
+			Parse_IntItem("/noDark", m_noDark);
+			continue;
+		}
+
 		// The Fit-window Settings
 		if(Equals(szToken, "FitWindow")){
 			ParseFitWindow();
+			continue;
+		}
+
+		// The Directory Mode Settings
+		if (Equals(szToken, "DirectoryMode")) {
+			ParseDirectoryMode();
 			continue;
 		}
 	}
@@ -462,5 +487,57 @@ int CMobileConfiguration::Parse_ShiftOrSqueeze(const CString &label, Evaluation:
 			option = Evaluation::SHIFT_LINK;
 		}
 	}
+	return 0;
+}
+
+
+/** Parses the settings for the DirectoryMode */
+int CMobileConfiguration::ParseDirectoryMode() {
+
+	// the actual reading loop
+	while (szToken = NextToken()) {
+		// Directory to watch for STD files in
+		if (Equals(szToken, "directory")) {
+			Parse_StringItem("/directory", m_directory);
+			continue;
+		}
+
+		// Spectrometer range
+		if (Equals(szToken, "spectrometerDynamicRange")) {
+			Parse_LongItem("/spectrometerDynamicRange", m_spectrometerDyanmicRange);
+			continue;
+		}
+
+		// Sleep time in ms between directory check
+		if (Equals(szToken, "sleep")) {
+			Parse_IntItem("/sleep", m_sleep);
+			continue;
+		}
+
+		// Default sky file
+		if (Equals(szToken, "defaultSkyFile")) {
+			Parse_StringItem("/defaultSkyFile", m_defaultSkyFile);
+			continue;
+		}
+
+		// Default dark file
+		if (Equals(szToken, "defaultDarkFile")) {
+			Parse_StringItem("/defaultDarkFile", m_defaultDarkFile);
+			continue;
+		}
+
+		// Default darkcur file
+		if (Equals(szToken, "defaultDarkcurFile")) {
+			Parse_StringItem("/defaultDarkcurFile", m_defaultDarkcurFile);
+			continue;
+		}
+
+		// Default sky file
+		if (Equals(szToken, "defaultOffsetFile")) {
+			Parse_StringItem("/defaultOffsetFile", m_defaultOffsetFile);
+			continue;
+		}
+	}
+
 	return 0;
 }