+++ DMSpecView.cpp
@@ -28,10 +28,12 @@
 #include "Configuration/Configure_Evaluation.h"
 #include "Configuration/Configure_GPS.h"
 #include "Configuration/Configure_Spectrometer.h"
+#include "Configuration/Configure_Directory.h"
 
 #include "MeasurementModes/Measurement_Traverse.h"
 #include "MeasurementModes/Measurement_Wind.h"
 #include "MeasurementModes/Measurement_View.h"
+#include "MeasurementModes/Measurement_Directory.h"
 #include "MeasurementModes/Measurement_Calibrate.h"
 #include <algorithm>
 #include <Mmsystem.h>	// used for PlaySound
@@ -66,6 +68,7 @@ BEGIN_MESSAGE_MAP(CDMSpecView, CFormView)
 
 	// Just view the spectra from the spectrometer without evaluations
 	ON_COMMAND(ID_CONTROL_VIEWSPECTRAFROMSPECTROMETER,	OnControlViewSpectra) // <-- view the output from the spectrometer
+	ON_COMMAND(ID_CONTROL_VIEWSPECTRAFROMDIRECTORY, OnControlProcessSpectraFromDirectory) // <-- view latest spectra file in directory
 
 	// Calibrate a spectrometer
 	ON_COMMAND(ID_CONTROL_CALIBRATESPECTROMETER,		OnControlCalibrateSpectrometer)
@@ -425,7 +428,7 @@ LRESULT CDMSpecView::OnDrawColumn(WPARAM wParam, LPARAM lParam){
 	m_ColumnPlot.SetSecondRange(0.0, 200, 0, m_minSaturationRatio, m_maxSaturationRatio, 0);
 
 	// Draw the columns (don't change the scale again here...)
-	if(m_spectrometerMode == MODE_TRAVERSE){
+	if(m_spectrometerMode == MODE_TRAVERSE || m_spectrometerMode == MODE_DIRECTORY){
 		if(fitRegionNum == 1){
 			m_ColumnPlot.SetPlotColor(m_PlotColor[0]);
 			if(m_showErrorBar){
@@ -492,7 +495,9 @@ LRESULT CDMSpecView::OnDrawSpectrum(WPARAM wParam,LPARAM lParam)
 	m_ColumnPlot.CleanPlot();
 	
 	// set the ranges for the plot
-	if(m_spectrometerMode == MODE_VIEW || m_spectrometerMode == MODE_CALIBRATE){
+	if(m_spectrometerMode == MODE_VIEW 
+		|| m_spectrometerMode == MODE_DIRECTORY
+		|| m_spectrometerMode == MODE_CALIBRATE){
 		m_ColumnPlot.SetRange(0.0, 2048, 0, m_minSaturationRatio, m_maxSaturationRatio, 0);
 	}else{
 		m_ColumnPlot.SetSecondRange(0.0, 200, 0, m_minSaturationRatio, m_maxSaturationRatio, 0);
@@ -528,8 +533,13 @@ LRESULT CDMSpecView::OnShowIntTime(WPARAM wParam, LPARAM lParam){
 	expTime.Format("%d  ms",m_Spectrometer->RequestIntTime());
 	this->SetDlgItemText(IDC_INTTIME,expTime);
 
-	m_Spectrometer->GetNSpecAverage(averageInSpectrometer, averageInComputer);
-	nAverage.Format("%dx%d", averageInSpectrometer, averageInComputer);
+	if (m_spectrometerMode == MODE_DIRECTORY) {
+		nAverage.Format("%d", m_Spectrometer->m_totalSpecNum);
+	}
+	else {
+		m_Spectrometer->GetNSpecAverage(averageInSpectrometer, averageInComputer);
+		nAverage.Format("%dx%d", averageInSpectrometer, averageInComputer);
+	}
 	this->SetDlgItemText(IDC_SPECNO, nAverage);
 
 	// Update the legend
@@ -654,7 +664,7 @@ LRESULT CDMSpecView::OnReadGPS(WPARAM wParam, LPARAM lParam)
 	this->SetDlgItemText(IDC_NGPSSAT, nSat);
 
 
-	if (!m_Spectrometer->m_gps->GotContact())
+	if (!(m_spectrometerMode == MODE_DIRECTORY) &&!m_Spectrometer->m_gps->GotContact())
 	{
 		// If the communication with the GPS is broken (e.g. device unplugged)
 		COLORREF warning = RGB(255, 75, 75);
@@ -757,6 +767,15 @@ void CDMSpecView::OnControlStart()
 {
 	if(!s_spectrometerAcquisitionThreadIsRunning)
 	{
+
+		CString cfgFile = g_exePath + TEXT("cfg.xml");
+		std::unique_ptr<Configuration::CMobileConfiguration> conf;
+		conf.reset(new Configuration::CMobileConfiguration(cfgFile));
+		if (conf->m_spectrometerConnection == conf->CONNECTION_DIRECTORY) {
+			OnControlProcessSpectraFromDirectory();
+			return;
+		}
+
 		/* Check that the base name does not contain any illegal characters */
 		CString tmpStr;
 		this->GetDlgItemText(IDC_BASEEDIT, tmpStr);
@@ -859,6 +878,28 @@ void CDMSpecView::OnControlViewSpectra(){
 	}
 }
 
+/** Starts the viewing of latest spectra in a directory specified by config file. */
+void CDMSpecView::OnControlProcessSpectraFromDirectory() {
+
+	CDMSpecDoc* pDoc = GetDocument();
+	CMeasurement_Directory *spec = new CMeasurement_Directory();
+	this->m_Spectrometer = (CSpectrometer *)spec;
+
+	pSpecThread = AfxBeginThread(CollectSpectra, (LPVOID)(m_Spectrometer), THREAD_PRIORITY_LOWEST, 0, 0, NULL);
+	s_spectrometerAcquisitionThreadIsRunning = true;
+	m_spectrometerMode = MODE_DIRECTORY;
+
+	m_ColumnPlot.SetYUnits("Column [ppmm]");
+	m_ColumnPlot.SetSecondYUnit("Intensity [%]");
+	m_ColumnPlot.SetXUnits("Number");
+	m_ColumnPlot.EnableGridLinesX(false);
+	m_ColumnPlot.SetBackgroundColor(RGB(0, 0, 0));
+	m_ColumnPlot.SetGridColor(RGB(255, 255, 255));
+	m_ColumnPlot.SetPlotColor(m_PlotColor[0]);
+	m_ColumnPlot.SetRange(0, 200, 1, 0.0, 100.0, 1);
+	m_ColumnPlot.SetMinimumRangeX(200.0f);
+	m_ColumnPlot.SetSecondRange(0.0, 200, 0, 0.0, 100.0, 0);
+}
 
 /** Starts the viewing of spectra from the spectrometer, 
 		without saving or evaluating them. */
@@ -1027,7 +1068,9 @@ void CDMSpecView::DrawSpectrum()
 	}
 
 	// if we're using a normal measurement mode...
-	if(m_spectrometerMode == MODE_TRAVERSE || m_spectrometerMode == MODE_WIND){
+	if(m_spectrometerMode == MODE_TRAVERSE 
+		|| m_spectrometerMode == MODE_WIND 
+		|| m_spectrometerMode == MODE_DIRECTORY){
 
 		// Plot the spectrum
 		m_ColumnPlot.SetPlotColor(m_Spectrum0Color);
@@ -1038,8 +1081,9 @@ void CDMSpecView::DrawSpectrum()
 			return;
 		}else{
 			memcpy(spectrum2, m_Spectrometer->GetSpectrum(1), sizeof(double)*spectrumLength);
-			for(int k = 0; k < spectrumLength; ++k)
+			for (int k = 0; k < spectrumLength; ++k) {
 				spectrum2[k] *= dynRange_inv;
+			}
 
 			m_ColumnPlot.SetPlotColor(m_Spectrum1Color);
 			m_ColumnPlot.SetLineWidth(2);
@@ -1055,6 +1099,9 @@ void CDMSpecView::DrawSpectrum()
 		}
 		else {
 			// Plot slave spectrum
+			if (m_Spectrometer->m_NChannels < 2) {
+				return;
+			}
 			memcpy(spectrum2, m_Spectrometer->GetSpectrum(1), sizeof(double)*spectrumLength);
 			for (int k = 0; k < spectrumLength; ++k) {
 				spectrum2[k] *= dynRange_inv;
@@ -1112,10 +1159,15 @@ void CDMSpecView::OnConfigurationOperation()
 	m_EvalPage.Construct(IDD_CONFIGURE_EVALUATION);
 	m_EvalPage.m_conf = configuration;
 
+	Configuration::CConfigure_Directory m_DirectoryPage;
+	m_DirectoryPage.Construct(IDD_CONFIGURE_DIRECTORY);
+	m_DirectoryPage.m_conf = configuration;
+
 	// Add the pages once they have been constructed
 	confDlg.AddPage(&m_specPage);
 	confDlg.AddPage(&m_gpsPage);
 	confDlg.AddPage(&m_EvalPage);
+	confDlg.AddPage(&m_DirectoryPage);
 
 	// Open the configuration dialog
 	confDlg.DoModal();
@@ -1195,7 +1247,7 @@ LRESULT CDMSpecView::OnShowInformationDialog(WPARAM wParam,LPARAM lParam){
 
 	Dialogs::CInformationDialog infDlg;
 	if(wParam == DARK_DIALOG)
-		infDlg.informationString = "The offset level has dropped more than 20 counts since last dark spectrum was measured. Please cover the telescope to collect a new dark spectrum.";
+		infDlg.informationString = "The offset level has dropped more than 20 counts since last dark spectrum was measured. Please consider restarting the measurements when appropriate.";
 	else if(wParam == INVALID_GPS)
 		infDlg.informationString = "Lost contact with GPS-satellites, no valid GPS-data recived.";
 	else if(wParam == CHANGED_EXPOSURETIME)